+++ src/main/java/com/parkez/common/image/ImageService.java
@@ -1,7 +1,7 @@
 package com.parkez.common.image;
 
-import com.parkez.common.image.dto.ImageDeleteRequest;
-import com.parkez.common.image.dto.ImageUploadRequest;
+import com.parkez.common.image.dto.request.ImageDeleteRequest;
+import com.parkez.common.image.dto.request.ImageUploadRequest;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
