+++ src/main/java/com/parkez/common/image/dto/request/ImageUploadRequest.java
@@ -1,16 +1,15 @@
-package com.parkez.common.image.dto;
+package com.parkez.common.image.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.springframework.web.multipart.MultipartFile;
-
-import java.util.List;
 
 @Getter
 @NoArgsConstructor
-public class ImageDeleteRequest {
+@Schema(description = "이미지 업로드/수정 시 요청 DTO")
+public class ImageUploadRequest {
 
     @NotBlank(message = "targetType은 필수 입력값입니다.")
     @Schema(description = "이미지 업로드할 도메인 입력", example = "UserProfile")
@@ -20,7 +19,9 @@ public class ImageDeleteRequest {
     @Schema(description = "이미지 업로드할 도메인의 개별 Id 입력", example = "1")
     private Long targetId;
 
-    public ImageDeleteRequest(String targetType, Long targetId) {
+
+    @Builder
+    public ImageUploadRequest(String targetType, Long targetId) {
         this.targetType = targetType;
         this.targetId = targetId;
     }
